+++ bee-launcher/bee_conf.json
@@ -1,1 +1,1 @@
-{"pyro4-ns-port": 53120}
+{"pyro4-ns-port": 41267}
