+++ bee-launcher/bee_orc_ctl.py
@@ -7,7 +7,7 @@ from subprocess import Popen
 from bee_aws_launcher import BeeAWSLauncher 
 from bee_vm_launcher import BeeVMLauncher
 from bee_os_launcher import BeeOSLauncher
-from bee_cc_launcher import BeeCCLauncher
+from bee_charliecloud_launcher import BeeCharliecloudLauncher
 import boto3
 from threading import Thread
 from bee_task import BeeTask
@@ -39,9 +39,10 @@ class BeeLauncherDaemon(object):
             beetask = BeeOSLauncher(total_tasks + 1, beefile)
             self.__beetasks[beetask_name] = beetask
             return beetask
-        elif exec_target == 'bee_cc':
-            beetask = BeeCCLauncher(total_tasks + 1, beefile)
+        elif exec_target == 'bee_charliecloud':
+            beetask = BeeCharliecloudLauncher(total_tasks + 1, beefile)
             self.__beetasks[beetask_name] = beetask
+            print("HERE ORC create task cc", str(beetask))
             return beetask
         
     def launch_task(self, beetask):
@@ -141,7 +142,7 @@ def main():
     update_system_conf(open_port)
     #Pyro4.naming.startNSloop(port = open_port, hmac = os.getlogin())
     Popen(['python', '-m', 'Pyro4.naming', '-k', os.getlogin(), '-p', str(open_port)])
-    time.sleep(5)
+    time.sleep(1) #HERE was 5
     bldaemon = BeeLauncherDaemon()
     daemon = Pyro4.Daemon()
     bldaemon_uri = daemon.register(bldaemon)
